requirements.txt
Flask
langchain_openai
langchain_community
PyPDF2
langchain_qdrant
qdrant-client
python-dotenv
flask-cors